Recipes project DONE

+++ src/main/java/com/recipes/demo/presentation/RecipeController.java
@@ -7,6 +7,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
 import java.time.LocalDateTime;
@@ -47,9 +48,14 @@ public class RecipeController {
     }
 
     @PutMapping("/{id}")
-    public void updateRecipe(@Valid @RequestBody Recipe recipe, @PathVariable Long id) {
+    public void updateRecipe(@Valid @RequestBody Recipe recipe, @PathVariable Long id,
+                             @AuthenticationPrincipal UserDetails currentUser) {
         var currentRecipe = service.findRecipeById(id);
+        User user = userService.findUserByEmail(currentUser.getUsername()).get();
         if (currentRecipe.isPresent()) {
+            if (currentRecipe.get().getUser().getEmail() != user.getEmail()) {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Forbidden access");
+            }
             LocalDateTime now = LocalDateTime.now();
             recipe.setId(id);
             recipe.setDate(now);
@@ -66,17 +72,21 @@ public class RecipeController {
                                         @AuthenticationPrincipal UserDetails currentUser) {
         User user = userService.findUserByEmail(currentUser.getUsername()).get();
         recipe.setUser(user);
-//        Recipe newRecipe = new Recipe(recipe.getId(), recipe.getName(), recipe.getCategory(),
-//                recipe.getDescription(), recipe.getDate(), recipe.getIngredients(), recipe.getDirections(), user);
         Recipe recivedRecipe = service.saveRecipe(recipe);
         Long idx = recivedRecipe.getId();
         user.getRecipes().add(recivedRecipe);
         return Map.of("id", idx);
     }
     @DeleteMapping("/{id}")
-    public void deleteRecipe(@PathVariable Long id) {
+    public void deleteRecipe(@PathVariable Long id,
+                             @AuthenticationPrincipal UserDetails currentUser) {
+        User user = userService.findUserByEmail(currentUser.getUsername()).get();
         var recipe = service.findRecipeById(id);
         if (recipe.isPresent()) {
+            if ((recipe.get().getUser() != null)
+                    && (recipe.get().getUser().getEmail() != user.getEmail())) {
+                throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Forbidden access");
+            }
             service.deleteRecipe(id);
             throw new RecipeNoContentException("Deleted for id = " + id);
         }
@@ -84,13 +94,4 @@ public class RecipeController {
             throw new RecipeNotFoundException("Recipe not found for id = " + id);
         }
     }
-
-//    @PostMapping("/register")
-//    public void register(@RequestBody User user) {
-//        // input validation omitted for brevity
-//
-//        user.setPassword(encoder.encode(user.getPassword()));
-//
-//        userRepo.save(user);
-//    }
 }
